requirements.txt
astroid==2.3.3
attrs==19.3.0
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
filelock==3.0.12
idna==2.8
importlib-metadata==1.2.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.0.2
packaging==19.2
pluggy==0.13.1
py==1.8.0
pycodestyle==2.5.0
pydash==4.7.6
pylint==2.4.4
pyparsing==2.4.5
pytest==5.3.1
python-dotenv==0.10.3
requests==2.22.0
six==1.13.0
toml==0.10.0
tox==3.14.2
typed-ast==1.4.0
urllib3==1.25.7
virtualenv==16.7.8
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
